+++ src/css/_population-slider.scss
