+++ src/css/style.scss
@@ -5,7 +5,7 @@
 @use "legend";
 @use "search";
 @use "filter";
-@use "slider";
+@use "population-slider";
 
 @import url("https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&display=swap");
