+++ src/js/populationSlider.js
