+++ src/js/setUpSite.js
@@ -7,7 +7,7 @@ import setUpSearch from "./search";
 import setUpAbout from "./about";
 import setUpDetails from "./cityDetails";
 import { setUpFilter } from "./filter";
-import setUpSlider from "./slider";
+import setUpSlider from "./populationSlider";
 
 const BASE_LAYER = new TileLayer(
   "https://stamen-tiles-{s}.a.ssl.fastly.net/toner-lite/{z}/{x}/{y}{r}.{ext}",
